feature: Keep junctions from geometry remove.
This commit adds the options "geometry.remove.keep-junctions.explicit"
and "geometry.remove.keep-junctions.input-file" mirroring the same
feature that was provided for edges. It allows one to remove geometry
while specifying a deny list for junctions that are to be preserved.

I added this option because I had a road network built out of grid
pieces, and when I removed the geometry, the interior junctions remained
as desired. However the exterior junctions at the corners were removed,
which was not desired.